Adding s4u (#59)
* Added S4U client side

Currently just handles S4U to self within the client
Fixed all the enums because the encoding is a little looser than AD prefers

* Added authenticator checksum for Samba interop

* Added SPNego message wrap

* Interop expects doubly-encoded GSS outputs

* We should omit fractional seconds

* GSS requires application specific data in the checksum

* But delegation shouldn't actually be set if ther's no ticket

* Client shouldn't overwrite session keys while in use

GetServiceTicket will fail if the TGT is updated between the start and end of the method call because the session key changes. Copying and locking the overwrite keeps the session key local to the call.

* Fix versioning

* Making the build happy

+++ Kerberos.NET/Asn1/Experimental/AsnReader.Integer.cs
@@ -205,7 +205,7 @@ namespace System.Security.Cryptography.Asn1
 
             if (TryReadSignedInteger(sizeof(int), expectedTag, UniversalTagNumber.Integer, out long longValue))
             {
-                obj = longValue;
+                obj = (int)longValue;
                 value = (T)obj;
                 return true;
             }
